Refactored CLI commands and extracted them into a separate module in forc-debug plugin.
